Convert horse statistics from byte to double precision
Fixed critical truncation bug where stat growth was being lost due to
byte casting. Stats now properly preserve fractional values enabling
gradual progression over multiple races.

Changes:
- HorseStatistic: Actual, DominantPotential, RecessivePotential now double
- Horse entity: All stat properties (Speed/Stamina/Agility/Durability/Happiness) return double
- StatProgressionCalculator: GrowStat uses double parameters
- RaceRunHorse: InitialStamina changed to double for consistency
- BreedingExecutor: Updated stat generation to use double
- All calculators: Updated to accept double stat values
- Test suite: Updated 517 tests to use double precision

Impact:
- Stat growth now works correctly (0.80, 1.92 preserved vs truncated to 0, 1)
- Horses progress realistically toward genetic ceilings over career
- All 517 unit tests passing

+++ TripleDerby.Services.Breeding/BreedingExecutor.cs
@@ -167,8 +167,8 @@ public class BreedingExecutor(
         int punnettQuadrant = randomGenerator.Next(1, 5);
         int whichGeneToPick = randomGenerator.Next(1, 3);
 
-        byte dominantPotential;
-        byte recessivePotential;
+        double dominantPotential;
+        double recessivePotential;
 
         HorseStatistic sireStatistic = sireStats.Single(x => x.StatisticId == statistic);
         HorseStatistic damStatistic = damStats.Single(x => x.StatisticId == statistic);
@@ -214,9 +214,9 @@ public class BreedingExecutor(
         dominantPotential = MutatePotentialGene(dominantPotential);
         recessivePotential = MutatePotentialGene(recessivePotential);
 
-        int min = Math.Max(1, dominantPotential / 3);
-        int maxExclusive = Math.Max(min + 1, dominantPotential / 2 + 1);
-        byte actual = (byte)randomGenerator.Next(min, maxExclusive);
+        int min = Math.Max(1, (int)(dominantPotential / 3.0));
+        int maxExclusive = Math.Max(min + 1, (int)(dominantPotential / 2.0) + 1);
+        double actual = randomGenerator.Next(min, maxExclusive);
 
         var foalStatistic = new HorseStatistic
         {
@@ -229,11 +229,11 @@ public class BreedingExecutor(
         return foalStatistic;
     }
 
-    private byte MutatePotentialGene(byte potential)
+    private double MutatePotentialGene(double potential)
     {
-        byte mutationMultiplier = (byte)randomGenerator.Next(1, 101);
+        int mutationMultiplier = randomGenerator.Next(1, 101);
         int mutationLowerBound;
-        byte mutationUpperBound;
+        int mutationUpperBound;
 
         switch (mutationMultiplier)
         {
@@ -251,9 +251,9 @@ public class BreedingExecutor(
                 break;
         }
 
-        potential = (byte)(potential + randomGenerator.Next(mutationLowerBound, mutationUpperBound));
+        potential = potential + randomGenerator.Next(mutationLowerBound, mutationUpperBound);
 
-        potential = (byte)(potential is < 30 or > 95 ? 50 : potential);
+        potential = potential is < 30 or > 95 ? 50 : potential;
 
         return potential;
     }
